gjmöhmgnöh

+++ dental-clinic-system/backend/api/patient/patient.go
@@ -2,6 +2,8 @@ package patient
 
 import (
 	"dental-clinic-system/application/patientService"
+	"dental-clinic-system/application/userService"
+	"dental-clinic-system/helpers"
 	"dental-clinic-system/models"
 	"encoding/json"
 	"github.com/gorilla/mux"
@@ -9,16 +11,25 @@ import (
 	"strconv"
 )
 
-func NewPatientController(patientService patientService.PatientService) *PatientHandler {
-	return &PatientHandler{patientService: patientService}
+func NewPatientController(patientService patientService.PatientService, userService userService.UserService) *PatientHandler {
+	return &PatientHandler{patientService: patientService, userService: userService}
 }
 
 type PatientHandler struct {
 	patientService patientService.PatientService
+	userService    userService.UserService
 }
 
 func (h *PatientHandler) GetPatients(w http.ResponseWriter, r *http.Request) {
-	patients, err := h.patientService.GetPatients()
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusNotFound)
+		return
+	}
+
+	patients, err := h.patientService.GetPatients(user.ClinicID)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusNotFound)
 		return
@@ -38,10 +49,25 @@ func (h *PatientHandler) GetPatient(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	patient, err := h.patientService.GetPatient(uint(id))
+
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusNotFound)
 		return
 	}
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusNotFound)
+		return
+	}
+
+	if patient.ClinicID != user.ClinicID {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	err = json.NewEncoder(w).Encode(patient)
 	if err != nil {
 		return
@@ -55,6 +81,16 @@ func (h *PatientHandler) CreatePatient(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, "User not found", http.StatusUnauthorized)
+		return
+	}
+
+	patient.ClinicID = user.ClinicID
 	patient, err = h.patientService.CreatePatient(patient)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -73,6 +109,27 @@ func (h *PatientHandler) UpdatePatient(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, "User not found", http.StatusUnauthorized)
+		return
+	}
+
+	GetPatient, err := h.patientService.GetPatient(patient.ID)
+
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusNotFound)
+		return
+	}
+
+	if user.ClinicID != GetPatient.ClinicID {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	patient, err = h.patientService.UpdatePatient(patient)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -92,6 +149,27 @@ func (h *PatientHandler) DeletePatient(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, "Invalid patient ID", http.StatusBadRequest)
 		return
 	}
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, "User not found", http.StatusUnauthorized)
+		return
+	}
+
+	GetPatient, err := h.patientService.GetPatient(uint(id))
+
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusNotFound)
+		return
+	}
+
+	if user.ClinicID != GetPatient.ClinicID {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	err = h.patientService.DeletePatient(uint(id))
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusNotFound)
